Refactor validations into a module
